KT-45777: Don't collect members of an added Java class as changes
as it's enough to report the name of the added Java class as changed (we
also do that for added Kotlin classes and Kotlin/Java removed classes).

Side node: IncrementalJvmCache sets this property
`collectAllMembersForNewClass` to `true` to collect members of an added class
(https://github.com/JetBrains/kotlin/commit/a978c6be354aa87040558616da060be985c7a4e2).
However, when set to `false` it doesn't break
IncrementalCompilationBaseIT tests, so it confirms that this property is
not needed.